feat(卡尔曼滤波): 添加预测位置显示功能并优化神经网络轨迹初始化
- 在卡尔曼滤波器中添加获取预测位置的方法
- 实现预测位置的渲染功能（青色虚线框）
- 将神经网络轨迹初始化逻辑提取为独立方法
- 修复yolo-detector-filter中的类ID引用问题
- 添加显示预测位置的配置选项

+++ src/curve.cpp
@@ -7,7 +7,7 @@
 #include <iostream>
 #include <random>
 
-static std::mt19937 global_rng(std::chrono::high_resolution_clock::now().time_since_epoch().count());
+static std::mt19937 global_rng(static_cast<unsigned int>(std::chrono::high_resolution_clock::now().time_since_epoch().count()));
 static std::uniform_real_distribution<double> global_dist(0.0, 1.0);
 
 double easeOutSine(double x) {
